Removed java11 source folders since JDK-11 is the baseline now
