+++ libs/core/src/main/java/org/opensearch/common/collect/Map.java
@@ -32,70 +32,66 @@
 
 package org.opensearch.common.collect;
 
-import java.util.AbstractMap;
-import java.util.Collections;
-import java.util.HashMap;
-
 public class Map {
 
     /**
-     * Returns an unmodifiable map containing one mapping.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of() {
-        return Collections.emptyMap();
+        return java.util.Map.of();
     }
 
     /**
-     * Returns an unmodifiable map containing one mapping.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1) {
-        return Collections.singletonMap(k1, v1);
+        return java.util.Map.of(k1, v1);
     }
 
     /**
-     * Returns an unmodifiable map containing two mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2) {
-        return mapN(k1, v1, k2, v2);
+        return java.util.Map.of(k1, v1, k2, v2);
     }
 
     /**
-     * Returns an unmodifiable map containing three mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2, K k3, V v3) {
-        return mapN(k1, v1, k2, v2, k3, v3);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3);
     }
 
     /**
-     * Returns an unmodifiable map containing four mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2, K k3, V v3, K k4, V v4) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4);
     }
 
     /**
-     * Returns an unmodifiable map containing five mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2, K k3, V v3, K k4, V v4, K k5, V v5) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5);
     }
 
     /**
-     * Returns an unmodifiable map containing six mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2, K k3, V v3, K k4, V v4, K k5, V v5, K k6, V v6) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6);
     }
 
     /**
-     * Returns an unmodifiable map containing seven mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(K k1, V v1, K k2, V v2, K k3, V v3, K k4, V v4, K k5, V v5, K k6, V v6, K k7, V v7) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7);
     }
 
     /**
-     * Returns an unmodifiable map containing eight mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(
         K k1,
@@ -115,11 +111,11 @@ public class Map {
         K k8,
         V v8
     ) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8);
     }
 
     /**
-     * Returns an unmodifiable map containing nine mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(
         K k1,
@@ -141,11 +137,11 @@ public class Map {
         K k9,
         V v9
     ) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8, k9, v9);
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8, k9, v9);
     }
 
     /**
-     * Returns an unmodifiable map containing ten mappings.
+     * Delegates to the Java9 {@code Map.of()} method.
      */
     public static <K, V> java.util.Map<K, V> of(
         K k1,
@@ -169,68 +165,30 @@ public class Map {
         K k10,
         V v10
     ) {
-        return mapN(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8, k9, v9, k10, v10);
-    }
-
-    @SuppressWarnings("unchecked")
-    private static <K, V> java.util.Map<K, V> mapN(Object... objects) {
-        if (objects.length % 2 != 0) {
-            throw new IllegalStateException("Must provide an even number of arguments to Map::of method");
-        }
-        switch (objects.length) {
-            case 0:
-                return Map.of();
-            case 2:
-                return Map.of((K) objects[0], (V) objects[1]);
-            default:
-                HashMap<K, V> map = new HashMap<>();
-                for (int k = 0; k < objects.length / 2; k++) {
-                    map.put((K) objects[k * 2], (V) objects[k * 2 + 1]);
-                }
-                return Collections.unmodifiableMap(map);
-        }
+        return java.util.Map.of(k1, v1, k2, v2, k3, v3, k4, v4, k5, v5, k6, v6, k7, v7, k8, v8, k9, v9, k10, v10);
     }
 
     /**
-     * Returns an unmodifiable map containing keys and values extracted from the given entries.
-     *
-     * @param <K> the {@code Map}'s key type
-     * @param <V> the {@code Map}'s value type
-     * @param entries {@code Map.Entry}s containing the keys and values from which the map is populated
-     * @return a {@code Map} containing the specified mappings
+     * Delegates to the Java9 {@code Map.ofEntries()} method.
      */
     @SafeVarargs
+    @SuppressWarnings("varargs")
     public static <K, V> java.util.Map<K, V> ofEntries(java.util.Map.Entry<? extends K, ? extends V>... entries) {
-        if (entries.length == 0) {
-            return Collections.emptyMap();
-        } else if (entries.length == 1) {
-            return Collections.singletonMap(entries[0].getKey(), entries[0].getValue());
-        } else {
-            HashMap<K, V> map = new HashMap<>();
-            for (java.util.Map.Entry<? extends K, ? extends V> entry : entries) {
-                map.put(entry.getKey(), entry.getValue());
-            }
-            return Collections.unmodifiableMap(map);
-        }
+        return java.util.Map.ofEntries(entries);
     }
 
     /**
-     * Returns an unmodifiable Map.Entry for the provided key and value.
+     * Delegates to the Java9 {@code Map.entry()} method.
      */
     public static <K, V> java.util.Map.Entry<K, V> entry(K k, V v) {
-        return new AbstractMap.SimpleImmutableEntry<>(k, v);
+        return java.util.Map.entry(k, v);
     }
 
     /**
-     * Returns an unmodifiable {@code Map} containing the entries of the given {@code Map}.
-     *
-     * @param <K> the {@code Map}'s key type
-     * @param <V> the {@code Map}'s value type
-     * @param map a {@code Map} from which entries are drawn, must be non-null
-     * @return a {@code Map} containing the entries of the given {@code Map}
+     * Delegates to the Java10 {@code Map.copyOf()} method.
      */
-    @SuppressWarnings({ "unchecked", "rawtypes" })
     public static <K, V> java.util.Map<K, V> copyOf(java.util.Map<? extends K, ? extends V> map) {
-        return (java.util.Map<K, V>) Map.ofEntries(map.entrySet().toArray(new java.util.Map.Entry[0]));
+        return java.util.Map.copyOf(map);
     }
+
 }
