+++ server/src/main/java/org/opensearch/monitor/jvm/JvmPid.java
@@ -32,31 +32,10 @@
 
 package org.opensearch.monitor.jvm;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.message.ParameterizedMessage;
-
-import java.lang.management.ManagementFactory;
-
 class JvmPid {
 
-    private static final long PID;
-
     static long getPid() {
-        return PID;
-    }
-
-    static {
-        PID = initializePid();
-    }
-
-    private static long initializePid() {
-        final String name = ManagementFactory.getRuntimeMXBean().getName();
-        try {
-            return Long.parseLong(name.split("@")[0]);
-        } catch (final NumberFormatException e) {
-            LogManager.getLogger(JvmPid.class).debug(new ParameterizedMessage("failed parsing PID from [{}]", name), e);
-            return -1;
-        }
+        return ProcessHandle.current().pid();
     }
 
 }
